[WIP] moved firestore sync to new jobs folder
